Disable form buttons when not available.
This adds user_info to the embedded javascript on the page to show each
user as logged in or out. It then disables the "IN" button for those
already clocked in, and the "OUT" button for those who are not clocked
in. If there is no user selected yet, then all buttons are disabled.