Timeclock CSS: be more explicit about layout
The existing CSS was not enforcing timeclock layout and so things were
getting misaligned depending on what the theme was doing with CSS, so
this updates the CSS to be more explicit on things that matter on the
timeclock page.

This shouldn't affect any pages other than the timeclock page.